refactor(lib): Split anubis.go up into some smaller specialized methods
